+++ pkgs/ffigen/lib/src/code_generator/func.dart
@@ -45,7 +45,7 @@ import 'writer.dart';
 class Func extends LookUpBinding with HasLocalScope {
   final FunctionType functionType;
   bool exposeSymbolAddress;
-  final bool exposeFunctionTypedefs;
+  bool generateTypedefs;
   bool isLeaf;
   final bool objCReturnsRetained;
   final bool useNameForLookup;
@@ -61,7 +61,7 @@ class Func extends LookUpBinding with HasLocalScope {
 
   bool get needsWrapper => !functionType.sameDartAndFfiDartType && !isInternal;
 
-  /// Contains typealias for function type if [exposeFunctionTypedefs] is true.
+  /// Contains typealias for function type if [generateTypedefs] is true.
   Typealias? _exposedFunctionTypealias;
 
   bool isIncluded = false;
@@ -77,7 +77,7 @@ class Func extends LookUpBinding with HasLocalScope {
     List<Parameter> parameters = const [],
     List<Parameter> varArgParameters = const [],
     this.exposeSymbolAddress = false,
-    this.exposeFunctionTypedefs = false,
+    this.generateTypedefs = false,
     this.isLeaf = false,
     this.objCReturnsRetained = false,
     this.useNameForLookup = false,
@@ -96,17 +96,19 @@ class Func extends LookUpBinding with HasLocalScope {
         functionType.parameters[i].symbol = Symbol('arg$i', SymbolKind.field);
       }
     }
+  }
 
-    // Get function name with first letter in upper case.
-    final upperCaseName = name[0].toUpperCase() + name.substring(1);
-    if (exposeFunctionTypedefs) {
-      _exposedFunctionTypealias = Typealias(
-        name: upperCaseName,
-        type: functionType,
-        genFfiDartType: true,
-        isInternal: true,
-      );
-    }
+  Typealias? fillExposedFunctionTypealias() {
+    if (!generateTypedefs) return null;
+    final upperCaseName = symbol.oldName.isEmpty
+        ? ''
+        : symbol.oldName[0].toUpperCase() + symbol.oldName.substring(1);
+    return _exposedFunctionTypealias ??= Typealias(
+      name: upperCaseName,
+      type: functionType,
+      genFfiDartType: true,
+      isInternal: true,
+    );
   }
 
   @override