+++ pkgs/ffigen/test/collision_tests/decl_symbol_address_collision_test.dart
@@ -27,7 +27,7 @@ void main() {
               enumClass: (node) => node.isIncluded = true,
               global: (node) => node.isIncluded = true,
               macroConstant: (node) => node.isIncluded = true,
-              typealias: (node) => node.isIncluded = true,
+              typealias: (node) => node.isIncluded = .always,
             ),
           ],
         ),
@@ -43,7 +43,7 @@ void main() {
             name: '_library',
             returnType: NativeType(SupportedNativeType.voidType),
             exposeSymbolAddress: true,
-            exposeFunctionTypedefs: true,
+            generateTypedefs: true,
           ),
           Func(
             name: '_SymbolAddresses_1',